+++ lib/page_flip_view.dart
@@ -4,3 +4,4 @@
 library;
 
 export 'src/page_flip_view.dart' show PageFlipController, PageFlipView;
+export 'src/page_spread.dart' show PageSpreadMode;