+++ lib/pdf.dart
@@ -2,4 +2,5 @@
 library;
 
 export 'src/page_flip_view.dart' show PageFlipController, PageFlipView;
+export 'src/page_spread.dart' show PageSpreadMode;
 export 'src/pdf/pdf_flip_book.dart';